refactor : features 내부에 있던 코드들을 pages 코드로 맥락에 맞게 변경
