+++ frontend/src/routes/betting_.$roomId.vote.voting.tsx
@@ -1,6 +1,6 @@
 import { createFileRoute, redirect } from "@tanstack/react-router";
-import { BettingPage } from "@/features/betting-page";
-import { getBettingRoomInfo } from "@/features/betting-page/api/getBettingRoomInfo";
+import { BettingPage } from "@/pages/betting-page";
+import { getBettingRoomInfo } from "@/pages/betting-page/api/getBettingRoomInfo";
 
 export const Route = createFileRoute("/betting_/$roomId/vote/voting")({
   component: BettingPage,
