+++ frontend/src/shared/lib/bettingRoomInfo.ts
@@ -1,5 +1,5 @@
 import { useSuspenseQuery } from "@tanstack/react-query";
-import { getBettingRoomInfo } from "@/features/betting-page/api/getBettingRoomInfo";
+import { getBettingRoomInfo } from "@/pages/betting-page/api/getBettingRoomInfo";
 
 export const bettingRoomQueryKey = (roomId: string) => ["bettingRoom", roomId];
